wip(details): scheduling routine project
