+++ src/app/pages/queue-entry/components/platform-list/platform-list.component.html
@@ -1,13 +1,16 @@
-<div>
+
+<div class="content">
   <mat-selection-list #platformlist [multiple]="false" (selectionChange)="onSelectChangeEvent($event)">
-    <mat-list-option *ngFor="let platform of platforms;" [value]="{ id: platform.id, value: platform.value}">
-      {{platform.value}}
+    <mat-list-option *ngFor="let platform of platforms;" [value]="{ id: platform.id, value: platform.value}" (click)="next()">
+      <div class="box-item-platform">
+        <img *ngIf="platform.value == 'PS4'" src="../../../../../assets/PS4.svg" alt="Ícone para games" srcset="">
+        <img *ngIf="platform.value == 'Sinuca/PingPong'" src="../../../../../assets/Ping-Pong.svg" alt="Ícone para games" srcset="">
+        <img *ngIf="platform.value == 'Monopoly' || platform.value == 'War'" src="../../../../../assets/snooker.svg" alt="Ícone para games" srcset="">
+        <div class="text-item">
+          <p>{{platform.value}}</p>
+          <p>4 Pessoas na fila</p>
+        </div>
+      </div>
     </mat-list-option>
   </mat-selection-list>
-
-  <ul *ngIf="platformQueue">
-    <li *ngFor="let queue of platformQueue;">
-      {{queue.queueCount}}
-    </li>
-  </ul>
 </div>
