+++ src/app/pages/queue-entry/components/queue-in/queue-in.component.scss
@@ -1,0 +1,32 @@
+.box-title {
+  display: flex;
+  align-items: center;
+
+  h1 {
+    margin: 0;
+  }
+}
+
+.list-users {
+  margin: 16px 0;
+  border: solid 1px #7a7a7a;
+  border-radius: 5px;
+
+  .item-user {
+    padding: 8px;
+    border-bottom: solid 1px #7a7a7a;
+
+    p {
+      margin: 0;
+    }
+  }
+
+  .item-user:first-child {
+    padding-top: 16px;
+  }
+
+  .item-user:last-child {
+    padding-bottom: 16px;
+    border: none;
+  }
+}
