+++ src/app/shared/components/header/header.component.html
@@ -1,3 +1,9 @@
-<div style="height: 80px; width: 100%; background: orangered;">
-
+<div class="header" style="">
+  <div class="icon">
+    <img src="../../../../assets/Fifo.svg" alt="Icone Fifo">
+    <p>ifo</p>
+  </div>
+  <span class="material-icons">
+    menu
+    </span>
 </div>
