+++ src/app/shared/components/header/header.component.scss
@@ -1,0 +1,23 @@
+.header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  height: 80px;
+  background: orangered;
+  padding: 0 16px;
+
+  .icon {
+    display: flex;
+
+    p {
+      color: #fff;
+      font-size: 42px;
+      margin: 20px 0 0 0;
+    }
+  }
+
+  span {
+    font-size: 42px;
+    color: #fff;
+  }
+}
