+++ src/app/shared/material/material.module.ts
@@ -13,6 +13,8 @@ import { MatListModule } from '@angular/material/list';
 import { MatTabsModule } from "@angular/material/tabs";
 import { MatInputModule } from '@angular/material/input';
 import {MatFormFieldModule} from '@angular/material/form-field';
+import {MatSnackBarModule} from '@angular/material/snack-bar';
+import {MatStepperModule} from '@angular/material/stepper';
 
 
 @NgModule({
@@ -31,7 +33,9 @@ import {MatFormFieldModule} from '@angular/material/form-field';
     MatListModule,
     MatTabsModule,
     MatInputModule,
-    MatFormFieldModule
+    MatFormFieldModule,
+    MatSnackBarModule,
+    MatStepperModule
   ]
 })
 export class MaterialModule { }
